:recycle: refactor: update unit test imports according to test-utils migration
